# api/gen/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    An example OpenAPI specification for the Little PM application, referencing path parameters from components.
  title: Little PM API
  version: 1.0.0
servers:
- description: Production server
  url: https://app.little.pm
...
paths:
...
  /tasks/{projectUUID}:
    get:
      description: Retrieve all tasks for the specified project.
      operationId: getProjectTasks
      parameters:
      - description: ID of the project
        explode: false
        in: path
        name: projectUUID
        required: true
        schema:
          $ref: '#/components/schemas/projectUUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Task'
                type: array
          description: A list of tasks
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Project not found
      summary: Get tasks for a project
      tags:
      - Tasks/Blueprint Service
      x-accepts:
      - application/json
      x-tags:
      - tag: Tasks/Blueprint Service
    post:
      description: Create a new task for the specified project.
      operationId: createTask
      parameters:
      - description: ID of the project
        explode: false
        in: path
        name: projectUUID
        required: true
        schema:
          $ref: '#/components/schemas/projectUUID'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Project not found
      summary: Create a new task
      tags:
      - Tasks/Blueprint Service
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Tasks/Blueprint Service
  /tasks/{projectUUID}/{taskUUID}:
    get:
      description: Retrieve a specific task by UUID within a project.
      operationId: getTask
      parameters:
      - description: ID of the project
        explode: false
        in: path
        name: projectUUID
        required: true
        schema:
          $ref: '#/components/schemas/projectUUID'
        style: simple
      - description: ID of the task
        explode: false
        in: path
        name: taskUUID
        required: true
        schema:
          $ref: '#/components/schemas/taskUUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task found
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Task or project not found
      summary: Get a task by UUID
      tags:
      - Tasks/Blueprint Service
      x-accepts:
      - application/json
      x-tags:
      - tag: Tasks/Blueprint Service
    patch:
      description: "Partially update a task’s status, completion date, etc."
      operationId: updateTask
      parameters:
      - description: ID of the project
        explode: false
        in: path
        name: projectUUID
        required: true
        schema:
          $ref: '#/components/schemas/projectUUID'
        style: simple
      - description: ID of the task
        explode: false
        in: path
        name: taskUUID
        required: true
        schema:
          $ref: '#/components/schemas/taskUUID'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateTask_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Task or project not found
      summary: "Update a task (e.g., completion)"
      tags:
      - Tasks/Blueprint Service
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Tasks/Blueprint Service
...
components:
...
  schemas:
...
    projectUUID:
      description: ID of the project
      example: 665c599d-5c8d-4d20-aaab-7ffaba150606
      format: uuid
      type: string
    taskUUID:
      description: ID of the task
      example: 665c599d-5c8d-4d20-aaab-7ffaba150606
      format: uuid
      type: string
...
    Error:
      description: Error object
      example:
        message: Error message
      properties:
        message:
          description: Error message
          example: Error message
          type: string
      type: object
...
    Task:
      example:
        name: name
        taskUUID: 665c599d-5c8d-4d20-aaab-7ffaba150606
        day: 0
        status: pending
      properties:
        taskUUID:
          description: ID of the task
          example: 665c599d-5c8d-4d20-aaab-7ffaba150606
          format: uuid
          type: string
        name:
          type: string
        day:
          type: integer
        status:
          enum:
          - pending
          - completed
          type: string
      type: object
    TaskCompletion:
      properties:
        taskUUID:
          description: ID of the task
          example: 665c599d-5c8d-4d20-aaab-7ffaba150606
          format: uuid
          type: string
        completedDate:
          format: date-time
          type: string
      type: object
...
    updateTask_request:
      oneOf:
      - $ref: '#/components/schemas/Task'
      - $ref: '#/components/schemas/TaskCompletion'
      x-one-of-name: UpdateTaskRequest
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT token
      scheme: bearer
      type: http